requirements.txt
Keras==2.2.0
tensorflow==1.8.0
numpy
pickle
nltk
pandas
Pillow==5.2.0
Flask==1.1.1
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Werkzeug==0.15.5
numpy>=1.9.2
scipy>=0.15.1
scikit-learn>=0.18
matplotlib>=1.4.3
pandas>=0.19
seaborn==0.9.0
wordcloud==1.7.0
tweepy==3.8.0
